+++ index.js
@@ -1,6 +1,0 @@
-module.exports = {
-    configure : require('./maps/bot-configurator'),
-    LocationResolvers : require('./maps/location-resolvers'),
-    MapProviders : require('./maps/map-providers'),
-    SearchFunctions: require('./maps/search-functions')
-}